requirements.txt
streamlit>=1.37
pandas>=2.0
PyYAML>=6.0
Pillow>=10.0